+++ backends-velox/src/main/scala/org/apache/gluten/execution/TopNTransformer.scala
@@ -16,8 +16,8 @@
  */
 package org.apache.gluten.execution
 
+import org.apache.gluten.execution.ValidationResult
 import org.apache.gluten.expression.ExpressionConverter
-import org.apache.gluten.extension.ValidationResult
 import org.apache.gluten.metrics.MetricsUpdater
 import org.apache.gluten.substrait.SubstraitContext
 import org.apache.gluten.substrait.rel.{RelBuilder, RelNode}
