+++ gluten-substrait/src/main/scala/org/apache/spark/sql/execution/python/EvalPythonExecTransformer.scala
@@ -18,8 +18,8 @@ package org.apache.spark.api.python
 
 import org.apache.gluten.backendsapi.BackendsApiManager
 import org.apache.gluten.execution.{TransformContext, TransformSupport, UnaryTransformSupport}
+import org.apache.gluten.execution.ValidationResult
 import org.apache.gluten.expression._
-import org.apache.gluten.extension.ValidationResult
 import org.apache.gluten.metrics.MetricsUpdater
 import org.apache.gluten.substrait.`type`._
 import org.apache.gluten.substrait.SubstraitContext
